+++ CHANGES.rst
@@ -14,4 +14,6 @@ Migrations
 Changes
 ~~~~~~~
 
+- Rename internal blue/wifi observation key to mac.
+
 - Removed migrations before version 1.4.